refactor _strip_base64_from_messages
